fix deploy error
