+++ .github/workflows/check.yaml
@@ -29,9 +29,6 @@ jobs:
       - run: npm install -g yarn
       - run: yarn install --immutable
 
-      - name: Build schema
-        run: yarn run build:schema
-
       - name: Run lint
         run: yarn run lint
