+++ .github/workflows/deploy-firebase-dev.yaml
@@ -57,9 +57,6 @@ jobs:
               }
             ]
 
-      - name: Build schema
-        run: yarn run 'build:schema'
-
       - name: Set firebase environment
         run: firebase use default --token=${{ secrets.FIREBASE_TOKEN }}
