+++ .github/workflows/deploy-firebase.yaml
@@ -57,9 +57,6 @@ jobs:
               }
             ]
 
-      - name: Build schema
-        run: yarn run 'build:schema'
-
       - name: Set firebase environment
         run: firebase use prod --token=${{ secrets.FIREBASE_TOKEN }}
