+++ .husky/pre-commit
@@ -1,6 +1,5 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
-yarn build:schema
 yarn fix
 yarn lint
