+++ projects/firestore/package.json
@@ -1,12 +1,12 @@
 {
   "name": "firestore",
+  "version": "0.0.1",
   "devDependencies": {
     "@firebase/testing": "^0.20.11",
     "@types/jest": "^28.1.3",
     "@types/node": "^18.0.0",
     "eslint": "^8.9.0",
     "jest": "^28.1.1",
-    "rnfb-bootstrap-schema": "workspace:*",
     "ts-jest": "^28.0.5",
     "ts-node": "^10.8.1",
     "typescript": "^4.7.4"
