+++ projects/functions/package.json
@@ -1,6 +1,6 @@
 {
   "name": "functions",
-  "private": true,
+  "version": "0.0.1",
   "engines": {
     "node": "16"
   },
@@ -32,7 +32,6 @@
     "eslint-plugin-import": "^2.25.4",
     "firebase-functions-test": "^0.2.0",
     "jest": "^29.3.1",
-    "rnfb-bootstrap-schema": "workspace:*",
     "ts-jest": "^29.0.3",
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4"
