+++ projects/functions/src/hello/service.ts
@@ -1,17 +1,15 @@
 import { firestore } from "firebase-admin";
-import { Firestore } from "rnfb-bootstrap-schema";
+
+import type { User } from "@common/firestore/schema";
+
 
 export const service = async (): Promise<boolean> => {
-  const user: Firestore.User.Schema = {
+  const user: User = {
     name: "bobby",
   };
   console.log("hello world", user);
 
-  const snapshot = await firestore()
-    .collection("users")
-    .withConverter(Firestore.User.Converter)
-    .doc("hoge")
-    .get();
+  const snapshot = await firestore().collection("users").doc("hoge").get();
   const data = snapshot.data();
   console.log("data exists", snapshot.exists);
   if (data !== undefined) {
