+++ projects/functions/src/utils/converter.ts
@@ -1,0 +1,3 @@
+export const convert = () => {
+  return {};
+};
