+++ projects/functions/tsconfig.json
@@ -8,13 +8,11 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"],
+      "@common/*": ["../../common/*"],
     }
   },
   "exclude": ["node_modules", "lib"],
   "include": [
     "./src"
-  ],
-  "references": [
-    {"path": "../schema"},
   ]
 }
