--- projects/schema/.eslintrc.json
@@ -1,6 +1,0 @@
-{
-  "extends": ["../../eslintrc.base.json"],
-  "ignorePatterns": [
-    "**/jest.config.js"
-  ]
-}
