--- projects/schema/README.md
@@ -1,1 +1,0 @@
-# schema
