--- projects/schema/jest.config.js
@@ -1,15 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  transform: {
-    "^.+\\.ts?$": "ts-jest",
-  },
-  transformIgnorePatterns: ["<rootDir>/node_modules/"],
-  moduleFileExtensions: ["js", "ts"],
-  globals: {
-    "ts-jest": {
-      tsconfig: "./tsconfig.json",
-    },
-  },
-  testMatch: ["**/*.test.ts"],
-};
