--- projects/schema/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "rnfb-bootstrap-schema",
-  "version": "0.0.1",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "scripts": {
-    "test": "jest",
-    "lint": "eslint --max-warnings=0 --fix --ext .ts .",
-    "fix": "yarn lint --fix",
-    "build": "tsc"
-  },
-  "files": [
-    "dist"
-  ],
-  "dependencies": {
-    "firebase": "^9.15.0",
-    "zod": "^3.20.2"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.2.5",
-    "@types/node": "^18.11.18",
-    "eslint": "^8.9.0",
-    "firebase-admin": "^11.4.1",
-    "jest": "^29.3.1",
-    "ts-jest": "^29.0.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.4"
-  }
-}
