--- projects/schema/src/firestore/index.ts
@@ -1,2 +1,0 @@
-export * as User from "./user";
-export * as Sample from "./sample";
