--- projects/schema/src/firestore/user.test.ts
@@ -1,12 +1,0 @@
-import { Converter } from "./user";
-
-describe("user", () => {
-  it("toFirestore", () => {
-    const data = Converter.toFirestore({
-      name: "nick",
-    });
-    expect(data).toStrictEqual({
-      name: "nick",
-    });
-  });
-});
