--- projects/schema/src/index.ts
@@ -1,1 +1,0 @@
-export * as Firestore from "./firestore";
