--- projects/schema/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "include": [
-    "src/**/*.ts"
-  ],
-  "exclude": ["node_modules"],
-  "compilerOptions": {
-    "skipDefaultLibCheck": true,
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "ts-node": {
-    "files": true
-  }
-}
